Remove runtime device/application license check for native app

+++ Rakefile
@@ -101,7 +101,8 @@ namespace "framework" do
 end
 
 
-$application_build_configs_keys = ['security_token', 'encrypt_database', 'android_title', 'iphone_db_in_approot', 'iphone_set_approot', 'iphone_userpath_in_approot', "motorola_license", "motorola_license_company","name", "iphone_use_new_ios7_status_bar_style", "iphone_full_screen"]
+$application_build_configs_keys = ['security_token', 'encrypt_database', 'android_title', 'iphone_db_in_approot', 'iphone_set_approot', 'iphone_userpath_in_approot', "name", "iphone_use_new_ios7_status_bar_style", "iphone_full_screen"]
+
 $winxpe_build = false
       
 def make_application_build_config_header_file
@@ -463,6 +464,11 @@ namespace "config" do
        $app_config[$config["platform"]]["capabilities"] and $app_config[$config["platform"]]["capabilities"].is_a? Array
     $app_config["capabilities"] = capabilities
 
+    # Add license related keys in case of shared runtime build
+    if $app_config['capabilities'].index('shared_runtime')
+      $application_build_configs_keys << 'motorola_license'
+      $application_build_configs_keys << 'motorola_license_company'
+    end
     application_build_configs = {}
 
     #Process rhoelements settings
@@ -619,7 +625,7 @@ namespace "config" do
         $rhoelements_features += "- Javascript API for device capabilities\n"                
     end
 
-    if File.exist?(File.join($app_path, "license.yml"))
+    if $app_config["capabilities"].index("shared_runtime") && File.exist?(File.join($app_path, "license.yml"))
         license_config = Jake.config(File.open(File.join($app_path, "license.yml")))    
     
         if ( license_config )
@@ -628,11 +634,7 @@ namespace "config" do
         end    
     end    
     
-    $invalid_license = false
-
-    if $rhoelements_features.length() > 0
-        $app_config['extensions'] << 'rhoelements-license' if $current_platform == "android"
-
+    if $app_config["capabilities"].index("shared_runtime") && $rhoelements_features.length() > 0
         #check for RhoElements gem and license
 	    if  !$app_config['re_buildstub']	
             begin
@@ -640,43 +642,16 @@ namespace "config" do
                 
                 $rhoelements_features = ""
                 
-                # check license
-                is_ET1 = (($current_platform == "android") and ($app_config["capabilities"].index("motorola")))
-                is_win_platform = (($current_platform == "wm") or ($current_platform == "win32") or $is_rho_simulator )
-
-                if (!is_ET1) and (!is_win_platform)
-                     # check the license parameter
-                     if (!$application_build_configs["motorola_license"]) or (!$application_build_configs["motorola_license_company"])
-                        $invalid_license = true
-                     end
-                end
-                
             rescue Exception => e
                 if $app_config['extensions'].index('nfc')
                     $app_config['extensions'].delete('nfc')
                 end
-                #if $app_config['extensions'].index('audiocapture')
-                #    $app_config['extensions'].delete('audiocapture')
-                #end
-                #if $app_config['extensions'].index('rho-javascript')
-                #    $app_config['extensions'].delete('rho-javascript')
-                #end
-                
                 if $application_build_configs['encrypt_database'] && $application_build_configs['encrypt_database'].to_s == '1'
                     $application_build_configs.delete('encrypt_database')
                 end
-                
-                #if $app_config['extensions'].index('signature')
-                #  $app_config['extensions'].delete('signature')
-                #end
             end
         end
     end
-
-    #if $app_config['extensions'].index('signature') && (($current_platform == 'iphone') || ($current_platform == 'android'))
-    #    $app_config['capabilities'] << 'signature'
-    #    $app_config['extensions'].delete('signature')
-    #end
 
     if $current_platform == "win32" && $winxpe_build == true
       $app_config['capabilities'] << 'winxpe'
@@ -688,11 +663,6 @@ namespace "config" do
     $app_config['extensions'].delete("mspec") if !$debug && $app_config['extensions'].index('mspec')
     $app_config['extensions'].delete("rhospec") if !$debug && $app_config['extensions'].index('rhospec')
     
-    if $invalid_license
-        $application_build_configs["motorola_license"] = '123' if !$application_build_configs["motorola_license"]
-        $application_build_configs["motorola_license_company"] = 'WRONG' if !$application_build_configs["motorola_license_company"]
-    end
-
     $rhologhostport = $config["log_host_port"] 
     $rhologhostport = 52363 unless $rhologhostport
     begin
@@ -2646,14 +2616,6 @@ at_exit do
     puts ' The following features are only available in RhoElements v2 and above:'
     puts $rhoelements_features
     puts ' For more information go to http://www.motorolasolutions.com/rhoelements '
-    puts '**************************************************************************************'
-  end
-  
-  if $invalid_license
-    puts '********* WARNING ************************************************************************'
-    puts ' License is required to run RhoElements application.'
-    puts ' Please, provide  "motorola_license" and "motorola_license_company" parameters in build.yml.'
-    puts ' For more information go to http://www.motorolasolutions.com/rhoelements '    
     puts '**************************************************************************************'
   end
